Move structs implementing request traits into db module
