+++ aquadoggo/src/graphql/client/mod.rs
@@ -2,10 +2,8 @@
 
 mod mutation;
 mod query;
-mod request;
 mod response;
 
 pub use mutation::ClientMutationRoot;
 pub use query::ClientRoot;
-pub use request::{EntryArgsRequest, PublishEntryRequest};
 pub use response::NextEntryArguments;
